+++ packages/frontend/src/mock/handlers/mockDir.ts
@@ -1,0 +1,3 @@
+import { BE_ORIGIN } from '~/constants';
+const mockDir = (path: string) => `${BE_ORIGIN}${path}`;
+export default mockDir;
